requirements.txt
Flask
gunicorn
flask-cors
apscheduler
python-dotenv
huggingface-hub
requests
